requirements.txt
transformers
datasets 
accelerate 
bitsandbytes 
xformers 
langchain 
sentence_transformers 
autotrain-advanced 
faiss-gpu
streamlit
PyMuPDF 
python-docx